# specs/002-blackbox-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tensol Blackbox MVP — Public HTTP API
  version: 0.1.0
  description: |
    REST contract between the Tensol backend (`server/`), the apps/site
    frontend, and (for the scan-complete webhook) the vps-agent running
    inside ephemeral GCP VMs.

    All routes versioned under `/v1`. All bodies validated with Zod on
    the server side per Constitution IX. All state-changing operations
    emit signed audit events per Constitution X. The HMAC-signed
    webhook contract is in `webhook.md`.
servers:
  - url: https://api.tensol.com
    description: Production
  - url: http://localhost:3001
    description: Local dev
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tensol_session
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: machine-readable code
          example: "free_quota_exhausted"
        message:
          type: string
          description: human-readable, safe to display
        retry_after_seconds:
          type: integer
          nullable: true
    ULID:
      type: string
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    Hostname:
      type: string
      maxLength: 253
      pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?(\\.[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?)*$"
      description: Lowercase RFC 1035 hostname, no trailing dot, no IP literal
    AttackSurfaceEntry:
      type: object
      required: [domain, primary, headers]
      properties:
        domain:
          $ref: "#/components/schemas/Hostname"
        primary:
          type: boolean
        headers:
          type: array
          maxItems: 10
          items:
            type: object
            required: [k, v]
            properties:
              k: { type: string, maxLength: 64 }
              v: { type: string, maxLength: 1024 }
    ScanOrder:
      type: object
      description: |
        Current pre-Convex REST shape. `payment_kind` and `amount_kopecks` are
        legacy pre-international-pivot fields; new billing should use
        provider-agnostic entitlements/credits rather than YooKassa-specific
        values.
      required: [id, user_id, status, tier, primary_domain, attack_surface,
                 safety_rps, payment_kind, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        user_id: { $ref: "#/components/schemas/ULID" }
        status:
          type: string
          enum: [draft, dns_pending, dns_verified, vm_provisioning,
                 running, completed, failed, cancelled]
        tier:
          type: string
          enum: [quick, deep]
        primary_domain: { $ref: "#/components/schemas/Hostname" }
        attack_surface:
          type: array
          items: { $ref: "#/components/schemas/AttackSurfaceEntry" }
        safety_rps:
          type: integer
          minimum: 1
          maximum: 500
        dns_verify_token: { type: string, nullable: true }
        dns_verified_at: { type: integer, nullable: true }
        scan_id:
          allOf: [{ $ref: "#/components/schemas/ULID" }]
          nullable: true
        failure_reason: { type: string, nullable: true }
        payment_kind:
          type: string
          deprecated: true
          description: Legacy compatibility value. New clients should use future entitlement/billing records.
          enum: [free_quick, yookassa]
        amount_kopecks:
          type: integer
          nullable: true
          deprecated: true
          description: Legacy RUB minor-unit value. New billing records should use amount_minor + ISO currency.
        created_at: { type: integer }
        updated_at: { type: integer }
    Scan:
      type: object
      required: [id, user_id, scan_order_id, profile, status, started_at]
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        user_id: { $ref: "#/components/schemas/ULID" }
        scan_order_id: { $ref: "#/components/schemas/ULID" }
        profile:
          type: string
          enum: [recon, standard, max]
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        started_at: { type: integer }
        completed_at: { type: integer, nullable: true }
        usage_tokens: { type: integer, nullable: true }
        usage_usd_cents: { type: integer, nullable: true }
...
paths:
...
  /v1/scan-orders:
    get:
      summary: List user's scan orders (dashboard)
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScanOrder" }
    post:
      summary: Create a draft scan order
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier, primary_domain]
              properties:
                tier:
                  type: string
                  enum: [quick]   # Deep doesn't create scan-orders in MVP
                primary_domain: { $ref: "#/components/schemas/Hostname" }
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOrder" }
        '429':
          description: Rate-limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/scan-orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    get:
      summary: Fetch one scan order
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOrder" }
        '404':
          description: Not found (also hides foreign-user existence)
    delete:
      summary: Cancel a scan order or scan in flight
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOrder" }
        '409':
          description: Cannot cancel in current state
  /v1/scan-orders/{id}/attack-surface:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    put:
      summary: Replace attack-surface list (Step 1 commit)
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attack_surface]
              properties:
                attack_surface:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items: { $ref: "#/components/schemas/AttackSurfaceEntry" }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOrder" }
        '409':
          description: Order is past the editable stage
  /v1/scan-orders/{id}/safety:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    put:
      summary: Set rate-limit (Step 2 commit)
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [safety_rps]
              properties:
                safety_rps:
                  type: integer
                  minimum: 1
                  maximum: 500
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOrder" }
  /v1/scan-orders/{id}/dns-verify/request:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    post:
      summary: Generate DNS verify token (Step 3 begin)
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Token returned
          content:
            application/json:
              schema:
                type: object
                required: [token, instructions]
                properties:
                  token: { type: string, example: "tensol-verify-01ARZ3..." }
                  instructions:
                    type: object
                    required: [record_type, record_name, record_value]
                    properties:
                      record_type: { type: string, enum: [TXT] }
                      record_name: { type: string, example: "@" }
                      record_value: { type: string }
                      ttl_hint: { type: integer, example: 300 }
  /v1/scan-orders/{id}/dns-verify/check:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    get:
      summary: Poll DNS verification status
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [verified, attempts]
                properties:
                  verified: { type: boolean }
                  attempts: { type: integer }
                  remaining_window_seconds: { type: integer }
                  last_error: { type: string, nullable: true }
  /v1/scan-orders/{id}/launch:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    post:
      summary: Launch the scan (Step 4 commit)
      tags: [scan-orders]
      security: [{ sessionCookie: [] }]
      responses:
        '202':
          description: Launched; scan record created
          content:
            application/json:
              schema:
                type: object
                required: [scan_id]
                properties:
                  scan_id: { $ref: "#/components/schemas/ULID" }
        '409':
          description: DNS not verified, or quota exhausted, or wrong status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        '429':
          description: Free quota exhausted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/scans/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ULID" }
    get:
      summary: Fetch one scan (used by Live + Findings pages, polled)
      tags: [scans]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
        '404':
          description: Not found
...